// config/types.ts
export interface IntegrationConfig {
  analytics:
    | { provider: "none" }
    | { provider: "google"; measurementId: string }
    | { provider: "pirsch"; code: string };
  ads:
    | { provider: "none" }
    | {
        provider: "adsterra-native";
        scriptUrl: string;
        containerId: string;
      };
  socialBar:
    | { provider: "none" }
    | {
        provider: "adsterra-social-bar";
        scriptUrl: string;
      };
  verification: {
    google: string | null;
    bing: string | null;
  };
}